sagemaker-cdk/requirements.txt
boto3
aws-cdk-lib==2.111.0
constructs>=10.0.0,<11.0.0
streamlit
requests
jinja2
matplotlib
sagemaker
ipywidgets
